pyproject.toml
[project]
name = "creditsystem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jupyter>=1.1.1",
    "pip>=26.0.1",
    "langchain>=0.1.0",
    "langchain-ollama>=0.1.0",
    "langchain-text-splitters>=0.0.1",
    "langchain-huggingface>=0.0.1",
    "langchain-chroma>=0.0.1",
    "langchain-google-genai>=0.1.0",
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
    "numpy>=1.26.0,<2.0.0",
    "scikit-learn>=1.3.0,<1.4.0",
    "litellm>=1.0.0",
    "gradio",
    "python-dotenv>=1.0.0",
]
